linting, minor debugging of mk_sam_file.py and mk_sam_utilities.py
Tried to weed out extraneous code and improve organization/formatting
(autopep8 was helpful in this regard).

I also spent some time testing the code on every old CSV file kept in
the Hargraves Dropbox (actually a really handy resource for testing any
future developments to the code). I was able to identify two minor bugs.
Both seemed to be pretty infrequent, but it's possible they are related
to recent issues with the code.

1) there was a section of the code where type was not explicitly
specified so that a few sample names would be read in as an int or float
rather than a string

2) the code would occasionally throw utf-8 encoding errors when reading
a csv file (no idea why this happens but I've experienced it before). I
just made it so that if this error was caught, the program switches to
an alternative encoding and attempts to reread the file rather than
aborting. This solved the issue every time.