check for correct sign of calculated mag dec
Make sure that calculated magnetic declination is sign-corrected. For
example, a declination of +350 should be output as -10.

Note that this sign-correction is merely a matter of convenience for
comparing declination values. Both values are functionally equivalent as
far as the calculations are concerned (probably why this went unnoticed
for so long).